fix(readArray): don't depend on host endianness
When the underlying AB bytes are transformed into a DataView it is done using the host
Endianness and not the user defined endianness.

This would work well for LE stored data and LE user system, but not when
they differ.

This PR tries to fix that case, and accounts for "uint8" and "int8"
arrays being endianness-independent according to my understanding.

The IIFE function to determine endianness was taken [from this
SO post](https://stackoverflow.com/questions/7869752/javascript-typed-arrays-and-endianness).

I added several tests that hopefully also spot any error in this
logic.